absolute CPU timing w/lag counter

+++ src/emu/mod.rs
@@ -1,7 +1,7 @@
 mod decode;
 mod chip;
 
-pub fn run(rom: Rom, mhz: f32) {
+pub fn run(rom: Rom, mhz: f64) {
     assert!(rom.rom.len() <= 0x8000, "exceeded maximum ROM size");
     chip::Chip::new(rom, mhz).run();
 }
